% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sav_labels.R
\name{add_sav_labels}
\alias{add_sav_labels}
\title{Add variable labels from a .sav file or labelled dataframe}
\usage{
add_sav_labels(data, match, from)
}
\arguments{
\item{data}{Dataframe to append a column of labels to.}

\item{match}{String indicating the name of the column in \code{data} that contains the variable names.}

\item{from}{Either a string indicating path to a \code{.sav} file or a "labelled" dataframe.}
}
\description{
Appends a column of variable labels, given a column of variable names and a \code{.sav} file
or labelled dataframe to get the labels from.
}
\examples{
# SPSS metadata will be blank, since it is an R dataset

library(haven)
write_sav(mtcars, "mtcars.sav")
labelled.df <- read_sav("mtcars.sav")

# using a path to .sav file
sack2::describe_df(labelled.df) \%>\%
 add_sav_labels(match = "variable", from = "mtcars.sav")

# using a labelled dataframe
sack2::describe_df(labelled.df) \%>\%
 add_sav_labels(match = "variable", from = labelled.df)
}
