% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charvec_tc.R
\name{charvec_tc}
\alias{charvec_tc}
\title{Copy character vector to clipboard as string of collapsed values for pasting}
\usage{
charvec_tc(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
A string of quoted elements of \code{x}, separated by
commas and linebreaks.
}
\description{
Takes a character vector, quotes each element and concatenates them, then
copies the collapsed string to clipboard.
}
\details{
Fast way to create code like \code{c("element1", "element2", ...)}
for pasting into code.
}
\examples{
charvec_tc(c("asdjalksjdlaksjd", "12312093u10923012", "asd"))
}
