% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charvec_tc_as_string.R
\name{charvec_tc_as_string}
\alias{charvec_tc_as_string}
\title{Send character vector to clipboard as string of collapse values for pasting}
\usage{
charvec_tc_as_string(x, width = 60)
}
\arguments{
\item{x}{Character vector}

\item{width}{Width at which to wrap}
}
\description{
Experimental. Addresses common situation in which you have a vector of
strings that you want to turn into a list like so
\code{c("element1", "element2", ...)} for pasting into code.
}
\examples{
charvec_tc_as_string(c("asdjalksjdlaksjd", "12312093u10923012", "asd"))
}
