% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charvec_to_casewhen_tc.R
\name{charvec_to_casewhen_tc}
\alias{charvec_to_casewhen_tc}
\title{Create a \code{case_when()} statement for a set of values and copy it to the clipboard}
\usage{
charvec_to_casewhen_tc(x, varname = "variable", sort = TRUE)
}
\arguments{
\item{x}{vector of values... there will be one \code{case_when()} line per unique value}

\item{varname}{string indicating name of variable for \code{case_when} statement}

\item{sort}{boolean. Should values in \code{x} be sorted before creating \code{case_when()} lines?}
}
\description{
Takes a vector of values, creates a corresponding \code{case_when()} statement
for those values, and copies that statement to the clipboard
}
\examples{
# charvec_to_casewhen_tc(c("northeast", "west", "south"),
#                        varname = "xvar")
#
# charvec_to_casewhen_tc(c("northeast", "west", "south"),
#                        varname = "xvar",
#                        sort = FALSE)
#
# charvec_to_casewhen_tc(c(1, 2, 3),
#                        varname = "xvar")
#
# charvec_to_casewhen_tc(c(1, 2, NA, 3),
#                        varname = "xvar")
#
# charvec_to_casewhen_tc(c(1, 2, NA, 3, 2),
#                        varname = "xvar")
}
