% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_output_dir.R
\name{clear_output_dir}
\alias{clear_output_dir}
\title{Delete any visible files in output/}
\usage{
clear_output_dir()
}
\description{
For use in standard repo structure.
}
\examples{
clear_output_dir()
}
