% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mice_distrib.R
\name{compare_mice_distrib}
\alias{compare_mice_distrib}
\title{Compare imputed vs. raw data for \code{mice} model}
\usage{
compare_mice_distrib(mice)
}
\arguments{
\item{mice}{\code{mids} object.}
}
\description{
Returns a dataframe with descriptive statistics for raw vs. imputed data.
}
\examples{
boys2 <- cbind(boys, dummy = c(NA, NA, NA, rbinom(nrow(boys) - 3, 1, .5)))
mice <- mice(boys2, method = c("pmm", "pmm", "pmm", "pmm", "pmm", "pmm", "polr", "pmm", "pmm", "logreg"))
}
