% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor0.R
\name{cor0}
\alias{cor0}
\title{Create tidy summary of pairwise correlations of vars in a dataframe}
\usage{
cor0(data)
}
\arguments{
\item{data}{Dataframe.}
}
\description{
Wraps typical workflow of \code{df \%>\% as.matrix() \%>\% Hmisc::rcorr() \%>\% broom::tidy()}.
}
\details{
Drops all factors with more than 2 levels.  Converts factors with 2 levels to numerics
via \code{as.numeric(x) - 1}.
}
\examples{
cor0(mtcars)
cor0(mice::boys)
}
