% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_all.R
\name{count_all}
\alias{count_all}
\title{Call \code{dplyr::count()} on all columns in a dataframe}
\usage{
count_all(data)
}
\arguments{
\item{data}{Dataframe.}
}
\description{
Useful for situations in which several columns are selected with regex, like
so: \code{df \%>\% select(matches("school")) \%>\% count_all()}.
}
\examples{
data <- mtcars[, c("cyl", "gear", "carb")]
count_all(data)
}
