% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_table.R
\name{cross_table}
\alias{cross_table}
\title{Version of \code{gmodels::CrossTable} with better defaults}
\usage{
cross_table(x, y, digits = 3, max.width = 5, expected = FALSE,
  prop.r = FALSE, prop.c = FALSE, prop.t = FALSE,
  prop.chisq = FALSE, chisq = FALSE, fisher = FALSE,
  mcnemar = FALSE, resid = FALSE, sresid = FALSE, asresid = FALSE,
  missing.include = FALSE, format = c("SAS", "SPSS"), dnn = NULL,
  ...)
}
\description{
Defaults to not including any of the proportions that clutter the table.
}
\details{
Copied from the \code{gmodels} source on 01/21/2017.

Use \code{?gmodels::CrossTable} to see arguments and examples.
}
\examples{
cross_table(mtcars$cyl, mtcars$gear)
}
