% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_NA.R
\name{describe_NA}
\alias{describe_NA}
\title{Return tidy summary of NAs in dataframe}
\usage{
describe_NA(data)
}
\arguments{
\item{data}{Dataframe.}
}
\description{
Function to return a dataframe with description of frequency of NAs
in target dataframe.
}
\examples{
describe_NA(mice::boys)
}
