% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_vec.R
\name{describe_vec}
\alias{describe_vec}
\title{Return tidy summary of vector}
\usage{
describe_vec(x, digits = NULL)
}
\arguments{
\item{x}{Vector.}

\item{digits}{Optional argument to round output to a specified number of digits.}
}
\description{
Function to return a dataframe with descriptives for a target vector
Mirrors \code{describe_df}, but for standalone vectors.
}
\details{
Note that the count in the \code{distinct} column excludes values of \code{NA}.
}
\examples{
describe_vec(mtcars$mpg)
describe_vec(mtcars$mpg, 2)
describe_vec(mice::boys$hc)

# handling of a factor
describe_vec(mice::boys$gen)

# NAs are excluded from the count of 'distinct'
describe_vec(c(0, 1, 1, 0, NA))

# a common use case, distinct from describe_df()
describe_vec(mtcars$hp / mtcars$wt)
}
