% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pdfs_in_dir.R
\name{extract_pdfs_in_dir}
\alias{extract_pdfs_in_dir}
\title{Gather all \code{.pdf} files in a directory, recursively}
\usage{
extract_pdfs_in_dir(dir)
}
\arguments{
\item{dir}{Directory to search for PDFs}
}
\description{
For use with Zotero, which exports PDFs of a batch of articles into
individual folders. Copies of all PDFs in directory will be placed into
a new folder at top of directory named \code{extracted-pdfs}.
}
\examples{
# non-operative
# extract_pdfs_in_dir(dir = "~/Downloads/syllabus-causalinference")
}
