% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnr_names.R
\name{fnr_names}
\alias{fnr_names}
\title{Find and replace patterns in names of a dataframe}
\usage{
fnr_names(data, pattern, replacement)
}
\arguments{
\item{data}{Dataframe.}

\item{pattern}{Regex pattern}

\item{replacement}{Regex replacement}
}
\description{
Designed for use in \code{magrittr} pipelines. Uses \code{stringr::str_replace}.
}
\examples{
df <- data.frame(idnum = c(1008, 1009, 1010, 1011),
                 height.t1 = c(42, 64, 12, 88),
                 height.t2 = c(50, 66, 15, 95))

df \%>\%
  fnr_names("\\\\.t", "\\\\.time") \%>\%
  fnr_names("height", "height.cm")
}
