% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnr_wd.R
\name{fnr_wd}
\alias{fnr_wd}
\title{Find and replace string in all \code{.R} files in directory}
\usage{
fnr_wd(existing.string, replacement.string)
}
\arguments{
\item{existing.string}{Pattern passed to \code{gsub}.}

\item{replacement.string}{Replacement passed to \code{gsub}.}
}
\description{
Only looks at \code{.R} files. Recursive.
}
\examples{
# non-operative
# fnr_wd("var1", "number.arrests")
# fnr_wd("rating", "teacher.rating)
}
