% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_NA.R
\name{get_num_NA}
\alias{get_num_NA}
\title{get_num_NA}
\usage{
get_num_NA(data)
}
\arguments{
\item{data}{Dataframe.}
}
\description{
Function to return a dataframe with description of freuqency of NAs
in target dataframe.
}
\examples{
get_num_NA(df)
}
