% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_train_perf.R
\name{get_train_perf}
\alias{get_train_perf}
\title{Get tidy training performance of a \code{caret::train} object}
\usage{
get_train_perf(trained.model)
}
\arguments{
\item{trained.model}{Result of a \code{train()} call in caret.}
}
\description{
Function to return a tidy dataframe with performance information
from a model trained by \code{caret::train}.
}
\details{
Writing this as a function that tidily gets performance (rather than tidies
performance returned by \code{getTrainPerf}) is necessary for access to the
\code{.$trainingData} for \code{nobs}.
}
\examples{
model1 <- caret::train(data = mtcars, mpg ~ ., method = "glm")
get_train_perf(model1)
}
