% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_sav.R
\name{inspect_sav}
\alias{inspect_sav}
\title{Inspect labeling metadata from .sav file}
\usage{
inspect_sav(path.or.df)
}
\arguments{
\item{path.or.df}{Either a string indicating location of .sav file or a "labelled" dataframe
already imported into R environment.}
}
\description{
Function that returns dataframe with name, label, values, and value
labels for all variables in an SPSS data file.
}
\examples{
# non-operative
# inspect_sav("my-data.sav")
# inspect_sav(my.df)
# inspect_sav(my.df, view = TRUE)
}
