% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key}
\alias{key}
\title{Create a random key}
\usage{
key(length = 5)
}
\arguments{
\item{length}{Length of key}
}
\description{
Create a random sequence of alphanumeric characters.
}
\examples{
key()
key(length = 10)
}
