% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_png.R
\name{load_png}
\alias{load_png}
\title{Load a .png file into graphics device}
\usage{
load_png(path)
}
\arguments{
\item{path}{Path to \code{.png} file.}
}
\description{
For interactive use, when checking that size parameters
produce usable graphic for a figure.
}
\examples{
# not run
# load_png("output/figure1.png")
}
