% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lp}
\alias{lp}
\title{Print tibble with all rows shown}
\usage{
lp(tibble)
}
\description{
*L*ong *p*rint to override standard tibble behavior.
}
\details{
Functionally, an alternative to \code{View}.
}
\examples{
tibble <- tibble::as_tibble(mtcars)

lp(tibble)

# cf. suppression behavior
print(tibble)
}
