% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lprint.R
\name{lprint}
\alias{lprint}
\title{Print all rows of a tibble}
\usage{
lprint(tibble)
}
\description{
"Long print" to override standard tibble behavior.
}
\details{
Functionally, an alternative to \code{View}.
}
\examples{
tibble <- tibble::as_tibble(mtcars)

lprint(tibble)

# cf. suppression behavior
print(tibble)
}
