% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_unique.R
\name{n_unique}
\alias{n_unique}
\title{n_unique}
\usage{
n_unique(vec)
}
\arguments{
\item{vector}{Vector.}
}
\description{
Function to count number of unique values in a vector.
}
\examples{
n_unique(vector)
}
