% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify_cell.R
\name{notify_cell}
\alias{notify_cell}
\title{Send a text notification to cell phone}
\usage{
notify_cell(cell.number, text.content = NULL)
}
\arguments{
\item{cell.number}{10-digit phone number to contact, without hyphens or periods.}

\item{text.content}{String that is body of message, i.e. contents of text.}
}
\description{
Send a text notification to cell phone.
Only works with AT&T carriers.
Uses MMS in order to appear as coming from email address to recipient, rather than a new number every time.
}
\examples{
# non-operative
# notify_cell(8005555555)
# notify_cell(8005555555, "The simulation is finished on laptop.")
}
