% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_number.R
\name{pad_number}
\alias{pad_number}
\title{Pad a number with leading zeroes}
\usage{
pad_number(x)
}
\arguments{
\item{x}{Numeric vector to pad with leading zeroes}
}
\description{
Pad a number with leading zeroes
}
\examples{
pad_number(c(1, 2, 12))
pad_number(c(1, 2, 120))

}
