% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecdf.R
\name{plot_ecdf}
\alias{plot_ecdf}
\title{Plot ECDF}
\usage{
plot_ecdf(data, var = NULL)
}
\arguments{
\item{data}{Either a vector, or a dataframe in which to find \code{var}.}

\item{var}{When \code{data} is a dataframe, bare or quoted name of column to use.}
}
\description{
Less typing to get \code{ggplot2::stat_ecdf}.
Can accept dataframe with bare or quoted name,
or vector.
}
\examples{
plot_ecdf(mtcars$mpg)
plot_ecdf(mtcars, mpg)
plot_ecdf(mtcars, "mpg")
}
