% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mice_distrib.R
\name{plot_mice_distrib}
\alias{plot_mice_distrib}
\title{plot_mice_distrib}
\usage{
plot_mice_distrib(mids.object, vars)
}
\arguments{
\item{mids.object}{\code{mids} object containing variables to plot.}

\item{vars}{Character vector of names of variables to plot.}
}
\description{
Plot imputed vs. raw data for \code{mice} models.
}
\details{
Will use the \code{mice} plotting methods for continuous variables,
and the \code{sack} custom plotting methods for categorical variables
(as indicated by the imputation method used).
}
\examples{
plot_mice_distrib(df.imp, c("cont.dv1", "dummy.dv2", "multicat.dv3"))
}
