% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mice_distrib_binary.R
\name{plot_mice_distrib_binary}
\alias{plot_mice_distrib_binary}
\title{Plot imputed vs. raw data for a binary variable in \code{mice} model}
\usage{
plot_mice_distrib_binary(mids.object, var.to.plot)
}
\arguments{
\item{mids.object}{\code{mids} object containing variables to plot.}

\item{var.to.plot}{String of name of the single variable to plot.}
}
\description{
Plot imputed vs. raw data for a single binary variable in \code{mice} models.
}
\examples{
plot_mice_distrib_binary(df.imp, "asian")
plot_mice_distrib_binary(df.imp, "ever.medicated")
}
