% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mice_distrib_poly.R
\name{plot_mice_distrib_poly}
\alias{plot_mice_distrib_poly}
\title{Plot imputed vs. raw data for a polytomous variable in \code{mice} model}
\usage{
plot_mice_distrib_poly(mids.object, var.to.plot)
}
\arguments{
\item{mids.object}{\code{mids} object containing variables to plot.}

\item{var.to.plot}{String of name of the single variable to plot.}
}
\description{
Plot imputed vs. raw data for a single multi-category variable in \code{mice} models.
}
\examples{
df.imp <- mice(boys, method = c("pmm", "pmm", "pmm", "pmm", "pmm", "pmm", "polr", "pmm", "pmm"))
plot_mice_distrib_poly(df.imp, "phb")
}
