% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_to_symbol.R
\name{pval_to_symbol}
\alias{pval_to_symbol}
\title{Convert p values to symbols}
\usage{
pval_to_symbol(p)
}
\arguments{
\item{p}{Vector of p values.}
}
\description{
Converts a vector of p values into character vector of the
symbols ***, **, *, †.
}
\examples{
pval_to_symbol(c(.50, .0004, .098, .044))
}
