% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round0.R
\name{round0}
\alias{round0}
\title{Round all numeric columns in a dataframe}
\usage{
round0(df, digits)
}
\arguments{
\item{digits}{Number of digits to round to.}

\item{data}{Dataframe whose numeric columns should be rounded.}
}
\description{
Round all numeric columns in a dataframe
}
\examples{
mtcars
round0(mtcars, 1)
}
