% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_session_info.R
\name{save_session_info}
\alias{save_session_info}
\title{Save session info as .txt file}
\usage{
save_session_info()
}
\description{
Will be named \code{session-info.txt}.
}
\examples{
save_session_info()
}
