% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_new_repo.R
\name{set_up_new_repo}
\alias{set_up_new_repo}
\title{Set up basic file structure for new repo}
\usage{
set_up_new_repo()
}
\description{
Creates \code{data/} and \code{output/} folders, plus \code{README.md},
\code{RUNME.R}, and \code{01-build-dataset.R}.
}
\examples{
set_up_new_repo()
}
