% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table0.R
\name{table0}
\alias{table0}
\title{Version of \code{table} with better defaults}
\usage{
table0(..., exclude = if (useNA == "no") c(NA, NaN), useNA = c("no",
  "ifany", "always"), dnn = list.names(...), deparse.level = 1)
}
\description{
First, \code{table0} is a version of \code{table()} in base R that automatically supplies \code{useNA = "always"}.
See that function for documentation. Second, \code{table1} is a wrapper for \code{table0} that automatically returns the proportions
(rather than frequencies) and rounds these to two digits.  Not sure how \code{table1} would work
with more than one dimension in table.
}
\examples{
table0()
table1()
}
