% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_as_comment.R
\name{tc_as_comment}
\alias{tc_as_comment}
\title{Write a string to the clipboard formatted as an 80-wide R comment}
\usage{
tc_as_comment(x)
}
\arguments{
\item{x}{string to be formatted into an R comment}
}
\description{
Wraps \code{x} to maximum line width of 80 and adds
\code{#} symbol at beginning of each line. Useful for long documentation.
}
\examples{
tc_as_comment("Call me Ishmael. Some years ago - never mind how long precisely -
having little or no money in my purse, and nothing particular to interest me on shore,
I thought I would sail about a little and see the watery part of the world. It
is a way I have of driving off the spleen, and regulating circulation.")
}
