% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_confusionMatrix.R
\name{tidy_confusionMatrix}
\alias{tidy_confusionMatrix}
\title{Tidy report of performance in \code{caret::confusionMatrix()}}
\usage{
tidy_confusionMatrix(confusionMatrix)
}
\arguments{
\item{confusionMatrix}{The result of a call to \code{caret::confusionMatrix()}}
}
\description{
Extracts all performance metrics embedded in the \code{confusionMatrix}, then removes
spaces and capitalization from names for ease in follow-up analyses.
}
\examples{

library(caret)

lvs <- c("normal", "abnormal")
truth <- factor(rep(lvs, times = c(86, 258)),
                levels = rev(lvs))
pred <- factor(
  c(
    rep(lvs, times = c(54, 32)),
    rep(lvs, times = c(27, 231))),
  levels = rev(lvs))

xtab <- table(pred, truth)

mat <- confusionMatrix(xtab)

print(mat)

tidy_confusionMatrix(mat)
}
