% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_MIcombine.R
\name{tidy_MIcombine}
\alias{tidy_MIcombine}
\title{Combine and tidy the result of a \code{mitools} analysis}
\usage{
tidy_MIcombine(results)
}
\arguments{
\item{results}{An \code{imputationResultList}, or list of results of an
analysis on a series of multiply imputed datasets.}
}
\description{
Produce a tidy dataframe of the results of a multiple imputation analysis
conducted with \code{mitools} package.
}
\examples{
library(mice)
library(mitools)

imp <- mice(boys)
imp.long <- complete(imp, action = "long")
imp.list <- imputationList(split(imp.long, imp.long$.imp))

results <- with(imp.list, glm(hgt ~ wgt))

tidy_MIcombine(results)
}
