% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_MIresults.R
\name{tidy_MIresults}
\alias{tidy_MIresults}
\title{Tidy the result of a \code{mitools} analysis}
\usage{
tidy_MIresults(MIresults)
}
\arguments{
\item{MIresults}{(i.e., the output of a \code{mitools::MIcombine} call).}
}
\description{
Produce a tidy dataframe of the results of a multiple imputation analysis conducted with \code{mitools} package.
}
\examples{
library(mice)
library(mitools)

imp <- mice(boys)
imp.long <- complete(imp, action = "long")
imp.list <- imputationList(split(imp.long, imp.long$.imp))

mi.results <- MIcombine(with(imp.list, glm(hgt ~ wgt)))

tidy_MIresults(mi.results)
}
