% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_missind.R
\name{to_missind}
\alias{to_missind}
\title{Recode a target column in dataframe to indicator of missingness}
\usage{
to_missind(data, x)
}
\arguments{
\item{data}{Dataframe}

\item{x}{String indicating name of target column}
}
\description{
Returns the input dataframe with the target column recoded such that
0 = missing and 1 = present.
}
\examples{
to_missind(mice::boys, "bmi")
}
