% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpsave.R
\name{wpsave}
\alias{wpsave}
\title{ggsave() wrapper that saves as .png+.pdf then loads to viewer}
\usage{
wpsave(.path, .height, .width)
}
\arguments{
\item{.path}{Path at which to save file}

\item{.height}{Height in inches}

\item{.width}{Width in inches}
}
\description{
Wraps \code{ggsave()}, but saves plot as both \code{.pdf}
and \code{.png}, then loads plot to viewer.
}
\examples{
# not run
# wpsave("output/figure1", .height = 4, .width = 4, )
}
