% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wp.R
\name{write_wp}
\alias{write_wp}
\title{Save dataframe as both .csv and .xlsx}
\usage{
write_wp(.data, .path)
}
\arguments{
\item{.data}{Dataframe to write to disk.}

\item{.path}{Path at which to save file. Note that extension
(e.g., \codes{.pdf}) should not be included.}
}
\description{
Convenience function to save a dataframe as both \code{.csv}
and \code{.xlsx}. Useful when working with others who prefer \code{.xlsx}.
}
\examples{
write_wp()
}
