% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_df.R
\name{describe_df}
\alias{describe_df}
\title{Return tidy summary of dataframe}
\usage{
describe_df(data)
}
\arguments{
\item{data}{Dataframe.}
}
\description{
Function to return a dataframe with descriptives for a target dataframe.
Each row in the returned dataframe will correspond to a column in the original dataframe.
}
\examples{
describe_df(mtcars)
describe_df(mice::boys)
}
