% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_df.R
\name{describe_df}
\alias{describe_df}
\title{Return tidy summary of dataframe}
\usage{
describe_df(data, digits = NULL)
}
\arguments{
\item{data}{Dataframe.}

\item{digits}{Optional argument to round output to a specified number of digits.}
}
\description{
Function to return a dataframe with descriptives for a target dataframe.
Each row in the returned dataframe will correspond to a column in the original dataframe.
}
\details{
Note that the count in the \code{distinct} column excludes values of \code{NA}.
}
\examples{
describe_df(mtcars)
describe_df(mtcars, 2)
describe_df(mice::boys)

# NAs are excluded from the count of 'distinct'
describe_df(data.frame(var1 = c(0, 1, 1, 0, NA)))
}
