% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mice_for_diag.R
\name{fit_mice_for_diag}
\alias{fit_mice_for_diag}
\title{Create a \code{mice} object for diagnostics}
\usage{
fit_mice_for_diag(..., iter10 = FALSE, Rhat.threshold = 1.1)
}
\arguments{
\item{...}{Arguments to be passed to \code{mice} call (typically, dataframe,
method vector, and prediction matrix).}

\item{iter10}{Logical switch indicating whether function should use the
iterative procedure to find a sufficient number of iterations. When
\code{TRUE}, will stop after fitting a \code{m = 5, maxit = 10} model.
Included for convenience when working with code that includes the function.}
}
\description{
Function that creates a \code{mice} object for diagnostics, with
\code{m = 5}. Iteratively increases the number of iterations by 5 until
Rhat < \code{Rhat.threshold} for mean and variance of all variables in imputation model.
Thus, accessing \code{$iteration} in the resulting \code{mice} object
will give the value of \code{maxit} to be used in the final \code{m > 5}
imputation model.
}
\details{
Always uses same random seed. Includes a dataframe with all Rhat values at
each step, which can be accessessed via \code{mice$rhat}. See \code{mice::mice}
for more documentation of arguments.
}
\examples{
library(mice)
pred.matrix <- quickpred(boys, mincor = .10, minpuc = 0.50)
mice.m5 <- fit_mice_for_diag(data = boys,
                             predictorMatrix = pred.matrix,
                             method = c("pmm"))
mice.m5$m
mice.m5$iteration
mice.m5$rhat
}
