% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mice_for_diag.R
\name{fit_mice_for_diag}
\alias{fit_mice_for_diag}
\title{Create a \code{mice} object for diagnostics}
\usage{
fit_mice_for_diag(..., iter10 = FALSE)
}
\arguments{
\item{...}{Arguments to be passed to \code{mice} call (typically, dataframe, method vector, and prediction matrix).}

\item{iter10}{Logical switch indicating whether function should use the iterative procedure
to find a sufficient number of iterations. When \code{TRUE}, will stop after fitting a \code{m = 5, maxit = 10} model. Included for convenience when working with code
that includes the function.}
}
\description{
Function that creates a \code{mice} object for diagnostics, with
\code{m = 5}. Iteratively increases the number of iterations by 5 until
Rhat < 1.05 for mean and variance of all variables in imputation model.
Thus, accessing \code{$iteration} in the resulting \code{mice} object
will give the value of \code{maxit} to be used in the final \code{m > 5}
imputation model.
}
\details{
See \code{mice::mice} for more documentation of arguments.
}
\examples{
library(mice)
pred.matrix <- quickpred(boys, mincor = .10, minpuc = 0.50)
df.imp.m5 <- fit_mice_for_diag(data = boys,
                               predictorMatrix = pred.matrix,
                               method = c("pmm", "pmm", "pmm", "pmm", "pmm", "pmm", "polr", "pmm", "pmm"))
df.imp.m5$m
df.imp.m5$iteration
}
