% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mice_for_diag.R
\name{fit_mice_for_diag}
\alias{fit_mice_for_diag}
\title{Create a \code{mice} object for diagnostics}
\usage{
fit_mice_for_diag(data, pred, method)
}
\arguments{
\item{data}{Dataframe to be imputed.}

\item{pred}{Square matrix indicating which variables should be used as predictors of which other variables.}

\item{method}{Character vector of methods to use for each variable in imputation model.}
}
\description{
Function that creates a \code{mice} object for diagnostics, with
\code{m = 5}. Iteratively increases the number of iterations by 5 until
Rhat < 1.05 for mean and variance of all variables in imputation model.
Thus accessing \code{$iteration} in the resulting \code{mice} object
will give the value of \code{maxit} to be used in the final \code{m > 5}
imputation model.
}
\details{
See \code{mice::mice} for more documentation of arguments.
}
\examples{
pred.matrix <- quickpred(boys, mincor = .10)
df.imp.m5 <- fit_mice_for_diag(boys, pred.matrix, method = c("pmm", "pmm", "pmm", "pmm", "pmm", "pmm", "polr", "pmm", "pmm"))
df.imp.m5$m
df.imp.m5$iteration
}
