% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varImp.R
\name{get_varImp}
\alias{get_varImp}
\title{Get tidy variable importance information out of model trained by \code{caret}}
\usage{
get_varImp(trained.model)
}
\arguments{
\item{trained.model}{Result of a \code{train()} call in caret.}
}
\description{
Function to return a tidy dataframe with variable importance information
from a model trained by \code{caret}.  Appends the algorithm and outcome information.
}
\examples{
model <- caret::train(data = mtcars, mpg ~ ., importance = TRUE)
get_varImp(model)
}
