% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varImp.R
\name{get_varImp}
\alias{get_varImp}
\title{Get tidy variable importance information out of model trained by \code{caret}}
\usage{
get_varImp(trained.model)
}
\arguments{
\item{trained.model}{Result of a \code{train()} call in caret.}
}
\description{
Function to return a tidy dataframe with variable importance information
from a model trained by \code{caret}.  Appends the algorithm and outcome information.
}
\details{
Note that an error will be thrown if the method used does not permit a \code{varImp} call,
that is, does not yield variable importance scores.  Currently shows a warning and returns
\code{NULL} whenever the call to \code{caret::varImp()} returns an error (e.g.,
if that algorithm has no variable importance metrics).
}
\examples{
model1 <- caret::train(data = mtcars, mpg ~ ., method = "glm")
get_varImp(model1)

# this one will show a warning and return null
model2 <- caret::train(data = mtcars, mpg ~ ., method = "ranger")
get_varImp(model2)
}
