% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_pred.matrix.R
\name{inspect_pred.matrix}
\alias{inspect_pred.matrix}
\title{Inspect the prediction matrix before imputation}
\usage{
inspect_pred.matrix(data, pred.matrix)
}
\arguments{
\item{data}{Dataframe that will be used in \code{mice} model.}
}
\description{
Function to check the prediction matrix before imputation.
Returns dataframe with following values for each variable in the model:
the number of missings, the number of variables used to impute it,
and the number of variables it is used to impute.
}
\examples{
pred.matrix = quickpred(boys, mincor = 0.10, minpuc = 0.50)
inspect_pred.matrix(boys, pred.matrix)
}
