% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_sav.R
\name{inspect_sav}
\alias{inspect_sav}
\title{Inspect labeling metadata from .sav file}
\usage{
inspect_sav(path.or.data, describe = FALSE)
}
\arguments{
\item{path.or.data}{Either a string indicating location of .sav file or a "labelled" dataframe
already imported into R environment.}

\item{describe}{Logical indicating whether to append descriptive information on each variable,
in addition to the SPSS metadata. (Calls \code{describe_df()} if so).}
}
\description{
Function that returns dataframe with name, label, values, and value
labels for all variables in an SPSS data file.
}
\examples{
# SPSS metadata will be blank, since it is an R dataset
library(haven)
write_sav(mtcars, "mtcars.sav")
inspect_sav("mtcars.sav", describe = TRUE)
}
