% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_sav.R
\name{inspect_sav}
\alias{inspect_sav}
\title{Inspect labeling metadata from .sav file}
\usage{
inspect_sav(path.or.df, view = FALSE)
}
\arguments{
\item{path.or.df}{Either a string indicating location of .sav file or a "labelled" dataframe
already imported into R environment.}

\item{view}{Logical indicating whether to return the dataframe or send to View().}
}
\description{
Function that returns dataframe with name, label, values, and value
labels for all variables in an SPSS data file.
}
\examples{
inspect_sav("my-data.sav")
inspect_sav(my.df)
inspect_sav(my.df, view = TRUE)
}
