% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lp}
\alias{lp}
\title{Print dataframe with all rows shown}
\usage{
lp(x)
}
\arguments{
\item{x}{\code{data.frame()} or \code{tibble()}}
}
\description{
*L*ong *p*rint to override standard data.frame() and tibble() printing behavior.
}
\details{
Functionally, an alternative to \code{View}.
}
\examples{

x <- do.call(rbind, replicate(20, as.matrix(mtcars), simplify = FALSE))

x <- as.data.frame(x)

print(x)
lp(x)

x <- tibble::as_tibble(x)

print(x)
lp(x)

}
