% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify_cell.R
\name{notify_cell}
\alias{notify_cell}
\title{notify_cell}
\usage{
notify_cell(cell.number, text.content)
}
\arguments{
\item{cell.number}{10-digit phone number to contact, without hyphens or periods.}

\item{text.content}{String that is body of message, i.e. contents of text.}
}
\description{
Send a text notification to cell phone via text.
Only works with ATT carriers.
Uses MMS in order to appear as coming from email address to recipient, rather than a new number every time.
}
\examples{
notify_cell(8005555555, "The simulation is finished.")
}
