% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mice_convergence.R
\name{plot_mice_convergence}
\alias{plot_mice_convergence}
\title{Plot imputation convergence for \code{mice} model}
\usage{
plot_mice_convergence(mids, vars)
}
\arguments{
\item{mids}{result of call to \code{mice()}.}

\item{vars}{Character vector of names of variables to plot.}
}
\description{
Plot imputation convergence for \code{mice} model
}
\examples{
df.imp <- mice(boys, method = c("pmm", "pmm", "pmm", "pmm", "pmm", "pmm", "polr", "pmm", "pmm"))
plot_mice_convergence(df.imp, c("hgt", "wgt"))
}
