% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mice_convergence.R
\name{plot_mice_convergence}
\alias{plot_mice_convergence}
\title{plot_mice_convergence}
\usage{
plot_mice_convergence(mids.object, vars)
}
\arguments{
\item{mids.object}{\code{mids} object containing variables to plot.}

\item{vars}{Character vector of names of variables to plot.}
}
\description{
Plot imputation convergence for \code{mice} models.
}
\examples{
plot_mice_convergence(df.imp, c("dv1", "dv2", "dv3"))
}
