% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mice_distrib.R
\name{plot_mice_distrib}
\alias{plot_mice_distrib}
\title{Plot imputed vs. raw data for \code{mice} model}
\usage{
plot_mice_distrib(mids.object, vars)
}
\arguments{
\item{mids.object}{\code{mids} object containing variables to plot.}

\item{vars}{Character vector of names of variables to plot.}
}
\description{
Will use the \code{mice} plotting methods for continuous variables,
and the \code{sack2} custom plotting methods for categorical variables
(as indicated by the imputation method used).
}
\examples{
boys2 <- cbind(boys, dummy = rbinom(nrow(boys), 1, .5))
df.imp <- mice(boys2, method = c("pmm", "pmm", "pmm", "pmm", "pmm", "pmm", "polr", "pmm", "pmm", "logreg"))
plot_mice_distrib(df.imp, c("hgt", "dummy", "phb"))
}
