% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mice_distrib.R
\name{plot_mice_distrib}
\alias{plot_mice_distrib}
\title{Plot imputed vs. raw data for \code{mice} model}
\usage{
plot_mice_distrib(mice, varlist)
}
\arguments{
\item{mice}{\code{mids} object containing variables to plot.}

\item{varlist}{Character vector of names of variables to plot.}
}
\description{
Chooses the type of plot based on \code{method} used for each variable.
}
\examples{
boys2 <- cbind(boys, dummy = c(NA, NA, NA, rbinom(nrow(boys) - 3, 1, .5)))
mice <- mice(boys2, method = c("pmm", "pmm", "pmm", "pmm", "pmm", "pmm", "polr", "pmm", "pmm", "logreg"))
plot_mice_distrib(mice, c("hgt", "dummy", "phb"))
}
