% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round0.R
\name{round0}
\alias{round0}
\title{round0}
\usage{
round0(df, digits)
}
\arguments{
\item{digits}{Number of digits to round to.}

\item{data}{Dataframe whose numeric columns should be rounded.}
}
\description{
Rounds all numeric columns in a dataframe to specific number of digits.
}
\examples{
round0(df, 0)
round0()
}
