% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_all.R
\name{source_all}
\alias{source_all}
\title{Source all R scripts within a folder}
\usage{
source_all(path, grepstring = "\\\\.R", print.source = TRUE)
}
\arguments{
\item{folder}{Target folder with scripts to be sourced.}
}
\description{
irectly copied from miceadds package (https://github.com/cran/miceadds);
that package loaded too many other extraneous things.
}
\examples{
source_all("helpers/")
source_all()
}
