% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_lines.R
\name{source_lines}
\alias{source_lines}
\title{source_lines}
\usage{
source_lines(file, lines)
}
\arguments{
\item{file}{character string with the path to the file to source.}

\item{lines}{numeric vector of lines to source in \code{file}.}
}
\description{
Source specific lines in an R file.
}
\details{
From https://gist.github.com/christophergandrud/1eb4e095974204b12af9.
}
\examples{
source_lines("load-data.R", 1 : 20)
}
