% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_lines.R
\name{source_lines}
\alias{source_lines}
\title{Source specific lines in an R file}
\usage{
source_lines(file, start = NULL, end = NULL)
}
\arguments{
\item{file}{character string with the path to the file to source.}

\item{start}{row number of first line in \code{file} to source.}

\item{end}{row number of last line in \code{file} to source.}
}
\description{
Source specific lines in an R file
}
\examples{
# non-operative
# source_lines("load-data.R", start = 50)
}
