% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vals.R
\name{vals}
\alias{vals}
\title{Create tidy dataframe that summarizes frequency of values in a vector}
\usage{
vals(vector)
}
\description{
Akin to 1-dimensional \code{table} with a few differences:
(1) it returns as a dataframe, and (2) it includes the proportion of elements.
}
\details{
Calls \code{print} internally and returns dataframe invisibly.
}
\examples{
vals(mtcars$gear)

# with NAs
vals(mice::boys$gen)
}
