% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpsave.R
\name{wpsave}
\alias{wpsave}
\title{Wrap ggsave() to save as .png & .pdf, then load to viewer}
\usage{
wpsave(path, height, width, dpi)
}
\arguments{
\item{path}{Path at which to save file. Note that if file extension is
specified, it will be stripped.}

\item{height}{Height in inches.}

\item{width}{Width in inches.}

\item{dpi}{Dots per inch, used for \code{.png}}
}
\description{
Wraps \code{ggsave()}, but saves plot as both \code{.pdf}
and \code{.png}, then loads plot to viewer.
}
\examples{
# not run; all three will produce same output
# wpsave("output/figure1", height = 4, width = 4)
# wpsave("output/figure1", height = 4, width = 4, dpi = 1000)
}
